+++ src/utils.js
@@ -1,89 +1,155 @@
-require('babel-polyfill')
-const semver      = require('semver')
-const fs          = require('fs')
+const semver = require('semver')
+const fs = require('fs')
 // const _        = require('lodash')
-const path        = require('path')
-const _           = require('lodash')
-const yaml        = require('js-yaml')
-const shell       = require('shelljs')
-const spawnSync   = require('spawn-sync')
-const oneLine     = require('common-tags/lib/oneLine')
-const stripIndent = require('common-tags/lib/stripIndent')
-const scrolex     = require('scrolex').persistOpts({
+const path = require('path')
+const _ = require('lodash')
+const yaml = require('js-yaml')
+const shell = require('shelljs')
+// const spawnSync   = require('spawn-sync')
+// const oneLine     = require('common-tags/lib/oneLine')
+// const stripIndent = require('common-tags/lib/stripIndent')
+const scrolex = require('scrolex').persistOpts({
   announce             : true,
   addCommandAsComponent: true,
 })
+const utils = this
+const oneLine = require('common-tags/lib/oneLine')
+// const pad = require('pad')
+// const async = require('async')
 
 if (require.main === module) {
   scrolex.failure(`Please only used this module via require`)
   process.exit(1)
 }
 
-const utils = module.exports
-
-module.exports.preferLocalPackage = (args, filename, appDir, name, entry, version) => {
-  let localModulePackage
-  let absoluteEntry
-  try {
-    localModulePackage = require(`${appDir}/node_modules/${name}/package.json`)
-    absoluteEntry      = fs.realpathSync(`${appDir}/node_modules/${name}/${entry}`)
-  } catch (e) {
-    localModulePackage = {}
-    absoluteEntry      = false
+module.exports.formatCmd = function formatCmd (cmd, { runtime, cmdName }) {
+  let extraVolumes = ''
+  if (runtime.contentBuildDir.indexOf(runtime.projectDir) === -1) {
+    extraVolumes = `--volume ${runtime.contentBuildDir}:${runtime.contentBuildDir}`
   }
 
-  if (localModulePackage.version && absoluteEntry) {
-    if (filename === absoluteEntry) {
-      return { type: 'symlinked', version: localModulePackage.version }
-    } else {
-      // We're entering globally and replacing this with a local instance
-      const exe = args.shift()
-      for (const i in args) {
-        // Replace the current entry, e.g. /usr/local/frey/lib/cli.js with the local package
-        if (args[i] === filename) {
-          args[i] = absoluteEntry
-        }
+  // Replace dirs
+  cmd = cmd.replace(/\[lanyonDir]/g, runtime.lanyonDir)
+  cmd = cmd.replace(/\[contentBuildDir]/g, runtime.contentBuildDir)
+  cmd = cmd.replace(/\[projectDir]/g, runtime.projectDir)
+  cmd = cmd.replace(/\[cacheDir]/g, runtime.cacheDir)
+
+  // Replace all npms with their first-found full-path executables
+  const npmBins = {
+    'browser-sync': 'node_modules/browser-sync/bin/browser-sync.js',
+    'nodemon'     : 'node_modules/nodemon/bin/nodemon.js',
+    'webpack'     : 'node_modules/webpack/bin/webpack.js',
+    // 'imagemin'     : 'node_modules/imagemin-cli/cli.js',
+  }
+  for (const name in npmBins) {
+    const tests = [
+      `/srv/lanyon/${npmBins[name]}`,
+      `${runtime.npmRoot}/${npmBins[name]}`,
+      `${runtime.projectDir}/${npmBins[name]}`,
+      `${runtime.lanyonDir}/${npmBins[name]}`,
+      `${runtime.gitRoot}/${npmBins[name]}`,
+    ]
+
+    let found = false
+    tests.forEach(test => {
+      if (fs.existsSync(test)) {
+        npmBins[name] = test
+        found = true
+        return false // Stop looking on first hit
       }
-      spawnSync(exe, args, { stdio: 'inherit' })
-      process.exit(0)
-      // return { type: 'local', version: localModulePackage.version }
+    })
+
+    if (!found) {
+      throw new Error(`Cannot find dependency "${name}" in "${tests.join('", "')}"`)
     }
-  } else {
-    return { type: 'local', version: version }
+    const pat = new RegExp(`(\\s|^)\\[${name}\\](\\s|$)`)
+    cmd = cmd.replace(pat, `$1node ${npmBins[name]}$2`)
+
+    // let nodeBin = utils.dockerString('node', { extraArgs: extraVolumes, runtime })
+    // cmd = cmd.replace(pat, `$1${nodeBin} ${npmBins[name]}$2`)
   }
+
+  // cp -f ${runtime.projectDir}/Gemfile ${runtime.cacheDir}/Gemfile &&
+  let jekyllBin = utils.dockerString('jekyll', { extraArgs: extraVolumes, runtime })
+
+  // Replace shims
+  cmd = cmd.replace(/(\s|^)\[jekyll\](\s|$)/, `$1${jekyllBin}$2`)
+
+  return cmd
 }
 
-module.exports.dockerCmd = ({cacheDir, projectDir, lanyonVersion}, cmd, flags) => {
-  if (!flags) {
-    flags = ''
-  }
+module.exports.dockerString = function dockerString (cmd, { extraArgs, runtime }) {
+  let wantVersion = runtime.lanyonVersion
+  wantVersion = '0.0.109'
   return oneLine`
     docker run
-      ${flags}
-      --user $(id -u)
-      --workdir ${cacheDir}
-      --volume ${cacheDir}:${cacheDir}
-      --volume ${projectDir}:${projectDir}
-    kevinvz/lanyon:${lanyonVersion}
-    ${cmd}
+      --rm
+      -i
+      --workdir ${runtime.cacheDir}
+      --volume ${runtime.cacheDir}:${runtime.cacheDir}
+      --volume ${runtime.projectDir}:${runtime.projectDir}
+      --volume ${runtime.cacheDir}/srv-jekyll:/srv/jekyll
+      ${extraArgs}
+      kevinvz/lanyon:${wantVersion}
+      ${cmd}
   `
 }
 
+module.exports.runString = async function runString (cmd, { runtime, cmdName, origCmd, hookName }) {
+  const scrolexOpts = {
+    stdio                : 'pipe',
+    cwd                  : runtime.cacheDir,
+    fatal                : true,
+    components           : `lanyon>${cmdName}>${origCmd.split('--')[0].trim().replace('[', '').replace(']', '')}`,
+    addCommandAsComponent: false,
+  }
+
+  // Run Pre-Hooks
+  scrolex.stick(`Running pre${hookName} hooks (if any)`)
+  await utils.runhooks('pre', hookName, runtime)
+
+  scrolex.exe(cmd, scrolexOpts, async (err, out) => { // eslint-disable-line handle-callback-err
+    // Run Post-Hooks
+    scrolex.stick(`Done. Running post${hookName} hooks (if any)`)
+    await utils.runhooks('post', hookName, runtime)
+  })
+}
+
 module.exports.runhooks = async (order, cmdName, runtime) => {
+  let squashedHooks = utils.gethooks(order, cmdName, runtime)
+
+  if (!squashedHooks) {
+    return
+  }
+
+  return scrolex.exe(squashedHooks, {
+    cwd       : runtime.projectDir,
+    mode      : (process.env.SCROLEX_MODE || 'passthru'),
+    components: `lanyon>hooks>${order}${cmdName}`,
+  })
+}
+
+module.exports.gethooks = (order, cmdName, runtime) => {
   let arr = []
 
   arr = [
     `${order}${cmdName}`,
     `${order}${cmdName}:production`,
     `${order}${cmdName}:development`,
+    `${order}${cmdName}:content`,
+    `${order}${cmdName}:content:production`,
+    `${order}${cmdName}:content:development`,
+    `${order}${cmdName}:assets`,
+    `${order}${cmdName}:assets:production`,
+    `${order}${cmdName}:assets:development`,
   ]
 
-  const collectStdout = {}
+  let squashedHooks = ''
   for (let i in arr) {
     let hook = arr[i]
     if (runtime[hook]) {
       const lastPart = hook.split(':').pop()
-      let needEnv    = 'both'
+      let needEnv = 'both'
 
       if (lastPart === 'production') {
         needEnv = lastPart
@@ -93,19 +159,14 @@ module.exports.runhooks = async (order, cmdName, runtime) => {
       }
 
       if (needEnv === 'both' || runtime.lanyonEnv === needEnv) {
-        let squashedHooks = runtime[hook]
+        squashedHooks = runtime[hook]
         if (_.isArray(runtime[hook])) {
           squashedHooks = runtime[hook].join(' && ')
         }
-        collectStdout[hook] = await scrolex.exe(squashedHooks, {
-          cwd : runtime.projectDir,
-          mode: (process.env.SCROLEX_MODE || 'singlescroll'),
-        })
+        return squashedHooks
       }
     }
   }
-
-  return collectStdout
 }
 
 module.exports.upwardDirContaining = (find, cwd, not) => {
@@ -126,7 +187,7 @@ module.exports.upwardDirContaining = (find, cwd, not) => {
   return false
 }
 
-module.exports.initProject = ({assetsBuildDir, gitRoot, cacheDir, binDir}) => {
+module.exports.initProject = ({ assetsBuildDir, gitRoot, cacheDir, binDir }) => {
   if (!fs.existsSync(assetsBuildDir)) {
     shell.mkdir('-p', assetsBuildDir)
     shell.exec(`cd "${path.dirname(gitRoot)}" && git ignore "${path.relative(gitRoot, assetsBuildDir)}"`)
@@ -149,48 +210,20 @@ module.exports.writeConfig = (cfg) => {
   if (!fs.existsSync(`${cfg.runtime.cacheDir}/jekyll.lanyon_assets.yml`)) {
     fs.writeFileSync(`${cfg.runtime.cacheDir}/jekyll.lanyon_assets.yml`, '# this file should be overwritten by the Webpack AssetsPlugin', 'utf-8')
   }
-  utils.fsCopySync(`${cfg.runtime.lanyonDir}/Gemfile.lock`, `${cfg.runtime.cacheDir}/Gemfile.lock`)
   try {
     fs.writeFileSync(`${cfg.runtime.cacheDir}/jekyll.config.yml`, yaml.safeDump(cfg.jekyll), 'utf-8')
   } catch (e) {
-    console.error({jekyll: cfg.jekyll})
+    console.error({ jekyll: cfg.jekyll })
     throw new Error(`Unable to write above config to ${cfg.runtime.cacheDir}/jekyll.config.yml. ${e.message}`)
   }
   fs.writeFileSync(`${cfg.runtime.cacheDir}/nodemon.config.json`, JSON.stringify(cfg.nodemon, null, '  '), 'utf-8')
   fs.writeFileSync(`${cfg.runtime.cacheDir}/full-config-dump.json`, JSON.stringify(cfg, null, '  '), 'utf-8')
-  fs.writeFileSync(`${cfg.runtime.cacheDir}/browsersync.config.js`, `module.exports = require("${cfg.runtime.lanyonDir}/lib/config.js").browsersync`, 'utf-8')
-  fs.writeFileSync(`${cfg.runtime.cacheDir}/webpack.config.js`, `module.exports = require("${cfg.runtime.lanyonDir}/lib/config.js").webpack`, 'utf-8')
+  fs.writeFileSync(`${cfg.runtime.cacheDir}/browsersync.config.js`, `module.exports = require("${cfg.runtime.lanyonDir}/src/config.js").browsersync`, 'utf-8')
+  fs.writeFileSync(`${cfg.runtime.cacheDir}/webpack.config.js`, `module.exports = require("${cfg.runtime.lanyonDir}/src/config.js").webpack`, 'utf-8')
   fs.writeFileSync(cfg.runtime.recordsPath, JSON.stringify({}, null, '  '), 'utf-8')
-
-  let dBuf = stripIndent`
-    FROM ruby:2.3.3-alpine
-    RUN mkdir -p /jekyll
-    WORKDIR /jekyll
-    ENV BUNDLE_APP_CONFIG /jekyll
-    RUN { \\
-      echo '---'; \\
-      echo 'BUNDLE_PATH: "/jekyll/vendor/bundler"'; \\
-      echo 'BUNDLE_DISABLE_SHARED_GEMS: "true"'; \\
-    } >> /jekyll/config
-    COPY Gemfile /jekyll/
-    COPY Gemfile.lock /jekyll/
-    RUN true \\
-      && apk --update add make gcc g++ \\
-      && (bundler install --force --path /jekyll/vendor/bundler || bundler update) \\
-      && rm -rf /var/cache/apk/* \\
-      && chmod 777 /jekyll/config \\
-      && true
-  `
-  fs.writeFileSync(`${cfg.runtime.cacheDir}/Dockerfile`, dBuf, 'utf-8')
-
-  let gBuf = `source 'http://rubygems.org'\n`
-  for (let name in cfg.runtime.gems) {
-    gBuf += `gem '${name}', '${cfg.runtime.gems[name]}'\n`
-  }
-  fs.writeFileSync(path.join(cfg.runtime.cacheDir, 'Gemfile'), gBuf, 'utf-8')
 }
 
-module.exports.satisfied = ({prerequisites, rubyProvidersSkip}, app, cmd, checkOn) => {
+module.exports.satisfied = ({ prerequisites }, app, cmd, checkOn) => {
   let tag = ''
   if (checkOn === undefined) {
     checkOn = app
@@ -198,24 +231,17 @@ module.exports.satisfied = ({prerequisites, rubyProvidersSkip}, app, cmd, checkO
     tag = `${checkOn}/`
   }
 
-  if (rubyProvidersSkip.indexOf(checkOn) !== -1) {
-    scrolex.failure(`${tag}${app} '${prerequisites[app].range} disabled via LANYON_SKIP`)
-    return false
-  }
-
   if (!cmd) {
     cmd = `${app} -v`
   }
 
-  const p              = shell.exec(cmd, { 'silent': true })
+  const p = shell.exec(cmd, { 'silent': true })
   const appVersionFull = p.stdout.trim() || p.stderr.trim()
-  const parts          = appVersionFull.replace(/0+(\d)/g, '$1').split(/[,p\s-]+/)
-  let appVersion       = parts[1]
+  const parts = appVersionFull.replace(/0+(\d)/g, '$1').split(/[,p\s-]+/)
+  let appVersion = parts[1]
 
   if (app === 'node') {
     appVersion = parts[0]
-  } else if (app === 'bundler') {
-    appVersion = parts[2]
   } else if (app === 'docker') {
     appVersion = parts[2]
   }